+++ eslint.config.js
@@ -30,6 +30,7 @@ export default [
       ...react.configs.recommended.rules,
       ...react.configs["jsx-runtime"].rules,
       ...reactHooks.configs.recommended.rules,
+      'react/prop-types': 'off',
       "react/jsx-no-target-blank": "off",
       "no-unused-vars": "warn",
       "react-refresh/only-export-components": [
