+++ src/App.jsx
@@ -1,19 +1,17 @@
 import { Route, Routes } from "react-router-dom";
-
-import AccountInform from "./layouts/AccountInform";
+import AccountInformation from "./layouts/AccountInformation";
 import Address from "./pages/AddressPage";
 import ChangePassword from "./pages/ChangePassword";
 import Profile from "./pages/ProfilePage";
-
 import ScrollToTop from "@/components/ScrollToTop";
 import ForgotPasswordLayout from "@/layouts/ForgotPasswordLayout";
 import MainLayout from "@/layouts/MainLayout";
 import About from "@/pages/AboutPage";
-import Contact from "@/pages/ContactPage";
+import Blog from "@/pages/Blog";
 import ForgotPassword from "@/pages/ForgotPassword";
 import Home from "@/pages/HomePage";
 import Login from "@/pages/LoginPage";
-import MyOrders from "@/pages/MyOrdersPage";
+import MyOrder from "@/pages/MyOrderPage";
 import OrderConfirmation from "@/pages/OrderConfirmationPage";
 import ProductDetails from "@/pages/ProductDetailsPage";
 import ProductLists from "@/pages/ProductListsPage";
@@ -25,24 +23,26 @@ import ResetPassword from "./pages/ResetPasswordPage";
 import ForgotPasswordVerify from "./pages/ForgotPassword/shared/ForgotPasswordVerify";
 import ThemeProvider from "./context/ThemeProvider";
 import ProductListsLayout from "./layouts/ProductListsLayout";
-import AddressListForUser from "./pages/AddressPage/shared/AddressListForUser";
-import AddressInform from "./pages/AddressPage/shared/AddressInform";
+import AddressInformation from "./pages/AddressPage/shared/AddressInformation";
+import AddressList from "./pages/AddressPage/shared/AddressList";
 
 const App = () => {
   return (
     <ThemeProvider>
       <ScrollToTop />
       <Routes>
-        {/* Route dependencies component Header and Footer */}
+        {/* Dependencies component Header and Footer */}
         <Route path="/" element={<MainLayout />}>
           <Route index element={<Home />} />
           <Route path="support" element={<Support />} />
-          <Route path="contact" element={<Contact />} />
+          <Route path="blog" element={<Blog />} />
           <Route path="about" element={<About />} />
         </Route>
 
+        {/* Login */}
         <Route path="/login" element={<Login />} />
 
+        {/* Forgot Password */}
         <Route path="/forgot-password" element={<ForgotPasswordLayout />}>
           <Route index element={<ForgotPassword />} />
           <Route
@@ -52,30 +52,34 @@ const App = () => {
           <Route path="reset-password" element={<ResetPassword />} />
         </Route>
 
+        {/* Register */}
         <Route path="/register" element={<Register />} />
         <Route path="/register/verify-account" element={<VerifyAccount />} />
 
+        {/* Prouduct Lists */}
         <Route path="/list-products" element={<ProductListsLayout />}>
           <Route index element={<ProductLists />} />
         </Route>
 
-        <Route path="/productDetails/:id" element={<ProductDetails />} />
-        <Route path="/shippingMethod" element={<ShippingMethod />} />
-        <Route path="/orderConfirmation" element={<OrderConfirmation />} />
+        {/* Product Details */}
+        <Route path="/product-details/:id" element={<ProductDetails />} />
 
-        {/* Route information user */}
-        <Route path="/accountInform" element={<AccountInform />}>
+        {/* Shipping Method */}
+        <Route path="/shipping-method" element={<ShippingMethod />} />
+
+        {/* Order Confirmation */}
+        <Route path="/order-confirmation" element={<OrderConfirmation />} />
+
+        {/* Account Inform */}
+        <Route path="/account-information" element={<AccountInformation />}>
           <Route path="profile/:id" element={<Profile />} />
-          <Route path="changePassword/:id" element={<ChangePassword />} />
+          <Route path="change-password/:id" element={<ChangePassword />} />
           <Route path="address/:id" element={<Address />} />
-          <Route
-            path="addressListForUser/:id"
-            element={<AddressListForUser />}
-          />
-          <Route path="newAddress/:id" element={<AddressInform />} />
+          <Route path="address-list/:id" element={<AddressList />} />
+          <Route path="new-address/:id" element={<AddressInformation />} />
         </Route>
 
-        <Route path="/myOrders" element={<MyOrders />} />
+        <Route path="/my-order" element={<MyOrder />} />
       </Routes>
     </ThemeProvider>
   );
