+++ src/layouts/AccountInformation/index.jsx
@@ -1,0 +1,17 @@
+import { Outlet } from "react-router-dom";
+
+import Footer from "../../components/Footer";
+import Header from "../../components/Header";
+import { Fragment } from "react";
+
+const AccountInformation = () => {
+  return (
+    <Fragment>
+      <Header />
+      <Outlet />
+      <Footer />
+    </Fragment>
+  );
+};
+
+export default AccountInformation;
