+++ src/layouts/MainLayout/index.jsx
@@ -1,7 +1,6 @@
 import { Outlet } from "react-router-dom";
-
-import Footer from "../../components/Footer";
-import Header from "../../components/Header";
+import Footer from "@/components/Footer";
+import Header from "@/components/Header";
 import { Fragment } from "react";
 
 export default function MainLayout() {
