+++ src/pages/AboutPage/index.module.css
@@ -1,12 +1,5 @@
-.content {
-    font-weight: 300;
-    color: var(--text-color);
-    margin: 12px 0;
-}
-
 .contentImg {
     width: 100%;
-    height: 450px;
     overflow: hidden;
 }
 
@@ -17,5 +10,5 @@
 }
 
 .contentImg img:hover {
-    transform: scale(1.2);
+    transform: scale(1.1);
 }
