+++ src/pages/Blog/index.jsx
@@ -1,0 +1,14 @@
+import WallpaperRepresentative from "@/components/WallpaperRepresentative";
+import { Box } from "@mui/material";
+
+const Blog = () => {
+  return (
+    <Box component={"section"}>
+      <WallpaperRepresentative titleHeader="Blog" />
+
+      
+    </Box>
+  );
+};
+
+export default Blog;
