+++ src/pages/HomePage/index.module.css
@@ -5,12 +5,12 @@
 
 .mediaImg {
     width: 300px;
-    height: 380px;
+    height: 400px;
 }
 
 .contentImg {
     width: 300px;
-    height: 380px;
+    height: 400px;
     position: absolute;
     top: 0;
     background-color: rgba(0, 0, 0, 0.4);
