+++ src/pages/LoginPage/index.jsx
@@ -257,7 +257,7 @@ const Login = () => {
                 borderRadius: 8,
                 objectFit: "cover",
               }}
-              src="/src/assets/images/background-fashions/background-form.jpg"
+              src="/src/assets/images/form/background-form.jpg"
               alt="Login background"
             />
           </Grid>
