+++ src/pages/MyOrdersPage/index.jsx
@@ -14,28 +14,35 @@ const MyOrders = () => {
 
   return (
     <Fragment>
-      <Header />
-      <WallpaperRepresentative titleHeader="Đơn hàng của tôi" />
-      <Container maxWidth="lg">
-        <Tabs
-          value={value}
-          onChange={handleChange}
-          aria-label="secondary tabs example"
-          variant="fullWidth"
-          sx={{
-            m: 4,
-            color: "white",
-            "& .MuiTab-root": { color: "var(--text-color)" },
-            // "& .Mui-selected": { color: "black" },
-            "& .MuiTabs-indicator": { backgroundColor: "black" },
-          }}
-        >
-          <Tab value="one" label="Chờ thanh toán" />
-          <Tab value="two" label="Đã thanh toán" />
-          <Tab value="three" label="Đã huỷ" />
-        </Tabs>
-      </Container>
-      <Footer />
+      <header>
+        <Header />
+      </header>
+      <main>
+        <WallpaperRepresentative titleHeader="Đơn hàng của tôi" />
+
+        <Container maxWidth="lg">
+          <Tabs
+            value={value}
+            onChange={handleChange}
+            aria-label="secondary tabs example"
+            variant="fullWidth"
+            sx={{
+              m: 4,
+              color: "white",
+              "& .MuiTab-root": { color: "var(--text-color)" },
+              // "& .Mui-selected": { color: "black" },
+              "& .MuiTabs-indicator": { backgroundColor: "black" },
+            }}
+          >
+            <Tab value="one" label="Chờ thanh toán" />
+            <Tab value="two" label="Đã thanh toán" />
+            <Tab value="three" label="Đã huỷ" />
+          </Tabs>
+        </Container>
+      </main>
+      <footer>
+        <Footer />
+      </footer>
     </Fragment>
   );
 };
