+++ src/pages/ProductDetailsPage/shared/ProductActions.jsx
@@ -173,7 +173,7 @@ const ProductActions = ({
     };
 
     dispatch(setOrderData(orderData));
-    navigate("/shippingMethod");
+    navigate("/shipping-method");
   };
 
   if (isCartLoading) return <Skeleton variant="rectangular" width={"100%"} height={30} />;
