+++ src/pages/RegisterPage/index.jsx
@@ -259,7 +259,7 @@ const Register = () => {
                 borderRadius: 8,
                 objectFit: "cover",
               }}
-              src="/src/assets/images/background-fashions/background-form.jpg"
+              src="/src/assets/images/form/background-form.jpg"
               alt="register background"
             />
           </Grid>
