+++ src/pages/ResetPasswordPage/index.jsx
@@ -265,7 +265,7 @@ const ResetPassword = () => {
                 borderRadius: 8,
                 objectFit: "cover",
               }}
-              src="/src/assets/images/background-fashions/background-form.jpg"
+              src="/src/assets/images/form/background-form.jpg"
             />
           </Grid>
         </Grid>
