+++ src/pages/SupportPage/index.jsx
@@ -1,51 +1,13 @@
-import SearchIcon from "@mui/icons-material/Search";
-import {
-  Box,
-  Container,
-  createTheme,
-  IconButton,
-  InputAdornment,
-  Stack,
-  TextField,
-  ThemeProvider,
-  useTheme,
-} from "@mui/material";
+import { Box, Container, createTheme, Stack } from "@mui/material";
 import { SimpleTreeView, TreeItem } from "@mui/x-tree-view";
 
 import WallpaperRepresentative from "@/components/WallpaperRepresentative";
 import "./index.css";
 
 const Support = () => {
-  const outerTheme = useTheme();
-
   return (
-    <section>
-      <WallpaperRepresentative
-        titleHeader="Chúng tôi có thể giúp gì cho bạn ?"
-        searchSupport={
-          <ThemeProvider theme={customTheme(outerTheme)}>
-            <TextField
-              id="outlined-basic"
-              label="Tìm kiếm sự hỗ trợ"
-              variant="standard"
-              sx={{
-                width: "660px",
-              }}
-              slotProps={{
-                input: {
-                  endAdornment: (
-                    <InputAdornment position="end">
-                      <IconButton edge="end">
-                        <SearchIcon sx={{ color: "white" }} />
-                      </IconButton>
-                    </InputAdornment>
-                  ),
-                },
-              }}
-            />
-          </ThemeProvider>
-        }
-      />
+    <Box component={"section"}>
+      <WallpaperRepresentative titleHeader="Support" />
 
       <Container maxWidth="lg">
         <Stack>
@@ -107,7 +69,7 @@ const Support = () => {
           </Box>
         </Stack>
       </Container>
-    </section>
+    </Box>
   );
 };
